requirements.lock.txt
# Deterministic runtime lock for core orchestration/evaluation tools
# Install with: pip install -r requirements.lock.txt
--extra-index-url https://download.pytorch.org/whl/cu121
fastapi==0.116.1
uvicorn==0.32.1
onnxruntime==1.22.1
onnx==1.18.0
numpy==1.26.4
pillow==11.1.0
python-multipart==0.0.20
opencv-python-headless==4.11.0.86
torch==2.5.1+cu121
torchvision==0.20.1+cu121
timm==1.0.26
scikit-learn>=1.3.0
tqdm>=4.66.0
matplotlib>=3.8.0
transformers==4.49.0
easyocr==1.7.1
paddlepaddle==3.3.1
paddleocr==2.7.0.3
ultralytics==8.3.0